creating the server part: part two

+++ src/controllers/card.ts
@@ -1,67 +1,78 @@
-import { Request, Response } from 'express';
+import { NextFunction, Request, Response } from 'express';
 import Cards from '../models/card';
+import { NotFoundError, BadRequst, Forbidden } from '../errors/index';
+import IRequest from '../utils/types/index';
 
-const createCard = (req: any, res: Response) => {
+const createCard = (req: IRequest, res: Response, next: NextFunction) => {
   const { name, link } = req.body;
-  const userId = req.user._id;
+  const userId = req.user?._id;
   Cards.create({ name, link, owner: userId })
     .then((card) => {
       if (!card) {
-        res.status(400).send('Переданы некорректные данные при создании карточки');
+        throw new NotFoundError('Нет пользователя с таким id');
       } else {
-        res.status(200).send(card);
+        res.status(201).send(card);
       }
     })
-    .catch(() => res.status(500).send('На сервере произошла ошибка'));
+    .catch(next);
 };
 
-const getCards = (req: Request, res: Response) => {
+const getCards = (req: Request, res: Response, next: NextFunction) => {
   Cards.find({})
     .then((cards) => res.send(cards))
-    .catch(() => res.status(500).send('На сервере произошла ошибка'));
+    .catch(next);
 };
 
-const deleteCard = (req: Request, res: Response) => {
+const deleteCard = (req: IRequest, res: Response, next: NextFunction) => {
   const { cardId } = req.params;
   Cards.findById(cardId)
     .then((card) => {
       if (!card) {
-        res.status(404).send('Карточка с указанным ID не найдена');
+        throw new NotFoundError('Карточка с указанным ID не найдена');
       } else {
-        Cards.findByIdAndDelete(req.params.cardId)
-          .then(() => res.status(200).send('Карточка успешна удалена'))
-          .catch((err) => res.status(500).send(`Произошла ошибка при удалении карточки: ${err}`));
+        const ownerId = card.owner.toString();
+        const userId = req.user?._id;
+        if (ownerId === userId) {
+          Cards.findByIdAndDelete(req.params.cardId)
+            .then(() => res.status(200).send('Карточка успешна удалена'))
+            .catch(next);
+        } else {
+          throw new Forbidden('Это не ваша карточка');
+        }
       }
     })
-    .catch(() => res.status(500).send('На сервере произошла ошибка'));
+    .catch(next);
 };
 
-const addLikes = (req: any, res: Response) => {
-  const userId = req.user._id;
+const addLikes = (req: IRequest, res: Response, next: NextFunction) => {
+  const userId = req.user?._id;
   const { cardId } = req.params;
   Cards.findByIdAndUpdate(cardId, { $addToSet: { likes: userId } }, { new: true })
     .then((card) => {
       if (!card) {
-        res.status(400).send('Переданы некорректные данные для постановки лайка');
+        throw new BadRequst('Переданы некорректные данные для постановки лайка');
       } else {
         res.status(200).send(card);
       }
     })
-    .catch(() => res.status(500).send('На сервере произошла ошибка'));
+    .catch(next);
 };
 
-const removeLikes = (req: any, res: Response) => {
-  const userId = req.user._id;
+const removeLikes = (req: IRequest, res: Response, next: NextFunction) => {
+  const userId = req.user?._id;
   const { cardId } = req.params;
+  if (!userId) {
+    throw new Error('Пользователь не аутентифицирован');
+  }
   Cards.findByIdAndUpdate(cardId, { $pull: { likes: userId } }, { new: true })
     .then((card) => {
       if (!card) {
-        res.status(400).send('Переданы некорректные данные для снятии лайка');
+        throw new BadRequst('Переданы некорректные данные для снятии лайка');
       } else {
         res.status(200).send(card);
       }
     })
-    .catch(() => res.status(500).send('На сервере произошла ошибка'));
+    .catch(next);
 };
 
 export {
